Introduced a new PlaygroundRuntime framework.
Unlike the playground transform runtime supplied by lldb, this framework provides concrete function values which can be set during initialization.
The intention is that separate frameworks like PlaygroundLogger will interact with the PlaygroundRuntime framework to register their entry points with the runtime.
Then, playground compilation simply needs to import this framework, and the playground transform will automatically insert calls to these function values.
Although this framework is named with "playground" in the name, it also provides runtime entry points for the PC macro.