Created a new PlaygroundLogger framework project.
This will be a brand new implementation of PlaygroundLogger, supporting the same stuff as the old library but hopefully in a more maintainable, testable manner.
This version of PlaygroundLogger is intended to work with the new PlaygroundRuntime framework, but entry points for use with legacy lldb will be provided as well. (Those are not yet present.)