[PlaygroundLogger] Introduced test targets for iOS and tvOS.
PlaygroundLogger needs to be tested on both iOS and tvOS, so the commit introduces test bundle and test host app targets for those platforms, adding to the test bundle already present for macOS.
It's important to note that, due to signing limitations, this testing only works in the simulator and not on real devices.
This commit also introduces iOS and tvOS PlaygroundLogger schemes which run the appropriate tests.