[PlaygroundLogger] Introduced `LogPolicy`, with support for capping the number of children logged.
The defaults are to log up to the first 10,000 children of "aggregates" (i.e. classes, structs, etc.) and up to the first 80 + last 20 of "containers" (i.e. arrays, dictionaries, sets, etc.).
This matches the behavior of the legacy PlaygroundLogger.
This commit includes introducing new tests that these policies are applied correctly.