[PlaygroundLogger] Added a workaround for issues dealing with nil `ImplicitlyUnwrappedOptional` values.
When an `ImplicitlyUnwrappedOptional` is bridged to Objective-C (such as via `instance as AnyObject`), the runtime "helpfully" unwraps it implicitly.
Unfortunately, that means if the `ImplicitlyUnwrappedOptional` is nil, the runtime crashes because it unexpectedly found nil.
This doesn't seem desirable in our case, as we're dealing with `Any` and not something that's explicitly `ImplicitlyUnwrappedOptional`.

As a result, make `ImplicitlyUnwrappedOptional` conform to a protocol which can indicate whether it's nil or not, and then check that before checking `-debugQuickLookObject`.

This allows `LogEntryTests.testNilIUO()` to pass.

This only affects swift-4.1-branch as `ImplicitlyUnwrappedOptional` has been removed on swift/master.